adds flow via eslint, integrates baseline eslint rules
The Flow/ESLint integration requires the use of the eslint-babel parser.  The most recent versions of the Vue ESLint plugin require the vue-eslint parser.  Result is that you can't have Flow ESLint and Vue ESLint working at the same time.  The older version of the Vue ESLint plugin doesn't require the vue-eslint parser, so it can work with Flow.  The tradeoff is that it really just allows for ESLint to run over the content of hte script tags, rather than analyzing the templates as well.