Date/Time Entry Fix (#1015)
As a front-end only solution the duplicate date time issue, I modified types.ts to take the DATE_TIME_LOWER_TYPE out of the TYPES_TO_LABEL dictionary, but added it as a EQUIV_TYPE to DATE_TIME_LOWER_TYPE, then I use EQUIV_TYPE in a new function, normalizedEquivalentType, which is used by TypeChangeMenu.vue's addMissingSuggestions and getSuggestedList function to map types to their supported equivalents. Additionally, I've added TIME_SUGGESTIONS to the guessTypeByValue function in types TS based on if Date.parse works on the value. Longer term, it'd be nice to have the types from the api as normalized as possible, but this is a suitable fix with flexibility for future inconsistent labeling if need to be in future.