Facet Menu Clean Up Fix (#1027)
Since we inject menus into facets, we have to handle the destruction of those elements, otherwise every time we update a facet, we inject a new instance of TypeChangeMenu while the old instances aren't cleaned up (even though they're no longer displayed, their instances are still trying to act on prop changes; they are basically orphaned by the underlying uncharted-software/facets component as it changes, which is why it needs to re-inject every time anyway.) To handle this, I'm now tracking the menus by their column name as they're created in facets.vue such that if we've already created a menu for that column name, we call destroy before binding a new menu to the location. We call destroy on all current menus as facets.vue is being destroyed since the destroy call on our the facets instance won't call destroy on the injected menus otherwise, which was resulting in continued attempts at acting on updates even on pages which had no facets on them.