Fixes to menu option suggestions.
Now menu suggestions are based only on the suggested types that are returned during the variables call for each dataset, such that an given type in the suggested types (passed from TypeChangeMenu to types.ts) is checked against a sugggestion dictionary, and the unique set of all possible types given the API suggestions is returned. It's definitely a bit overbroad, but it is very consistent and won't change the list of options based on the current selected type. Long term, we still need to basically do some assessments on intake where we do the kind of value parsing the front end was trying to help narrow down the likely possible types, but that's an after July thing. This will at least make the demos consistent. Also, caught one more error case in the action such that 400s on time series failures are also caught and populated in the facet. We could do the same for Geocoding failures, but I think that might get in the way of dev for now.