Add search to join suggestion panel
Noticed that when the service is up and we get join suggestions back, we get a lot of them, so to help folks narrow down the desired dataset even further, I have added a simple search filter in the panel that checks the dataset name and description for the query the user types and filters the list down on that basis. Added some extra checks on certain deep object traversals we make for for columns as that doesn't always return.