Bypass Join Column Selection For Direct Join #1109
Bypasses join selection for direct join after selecting a target join dataset in the join panel. Join button in panel attempts a join, if successful display preview modal where join can still be committed or canceled. All changes made are in service of making that work, including bypassing the hack for the actual datamarts as we don't have the desired column info in the hack.