Magnifying glass fix #1219
In the facet refactor, missed was the issue of the ImagePreview vue component attempts to find conditional dom elements immediately after state had been updated by the magnify action to inject additional dom elements via jQuery. I'm not entirely convinced this code even with this fix is ideal as it's basically trying to avoid an extra image load for the magnified image by copying a dom element that already contains the image into the modal. However, long term, I don't think we'd want the full images loading as the thumbnail sized views in the facets, we'd want actual thumbnails to save bandwidth/load time, then only load the full version when we open the modal.