Watching imageURL in for changes #1220
Due to some quirks in the caching optimizations made to image loading, the imageUrl property could change for a rendered image preview component instance without triggering a redraw of the underlying image, as seen in the defect. By watching the imageUrl prop, we can clear the existing image and request the image if it isn't yet loaded in the store or load it from the store.