Cut return feature when image selected #1216
Some code that for some reason always returned label with image code. It fixes the regression in the bug report, but I want to know why we had this code explicitly in here in the first place before any PRs. Kevin made the original change as part of time series stuff in March.