Facet Header Truncation & Tooltip #1261
This covers only the truncation and basic tooltip to display the untruncated title for facets. Since the the work for table header truncation touches different files, there's no reason to let it block getting this section in first. This doesn't yet cover any of the work in variable description story other than providing tooltip we could later fill in with more than the untruncated title.