#1343 - table updating geocoordinate fix
issue with the look up for variable descriptions used in table tooltips failing when, in the case of geocoordinate we're looking for a description for a column we've since combined into the geocoordinate. solution was to check the variable look up worked before trying to pull description out of the variable. only side effect is no description for that variable in the table, but this was okayed by Chris.