#1353 - fix regression type detection
regression enabled function upon which error bounding depends to even be displayed expected task type to only ever be a single item, not multiple (like "?type=regression,univariate",) so it broke. Fix is just to split up the return from the route task getter and then see if there's regression in that set.