#1351 - Call Refresh On WebSocket Close
added a callback to be called on socket close, which I then I populated in the call from the create solutions from view with a simple vue router refresh via the go function. It's not the the most elegant solution, but I should be fairly rebust to changes in the pipeline messaging as so long as the pipeline closes after we navigate to the results view, we're safe to refresh.