#1345 Filtering Reload Fix
Added a means of removing unused feature variables based on the incoming set of variables to be used such that we don't hard clear and reload everything, which in turn keeps filtering triggering that hard clear and reload. Meanwhile, I manually verified that switching between models still clears the unused variables should the newly selected model use different models for it's predictions.