#1369 Prediction Inference View
Initial commit that will display something when we load the prediction view. However, currently that is just the data from results/view model. Every function in the prediction store is just lifted from results for now. I'm getting this up in a branch for advance visibility/feedback as I've also some what aligned the prediction view with the design in the card as well.