#1369 Calls Prediction Endpoint Loads Data Into Table
Also has bunch of duplicate actions to either clear or rename out of the predictions' area of the store, but it does actually work technically now, just puts a bunch of warnings in the console about the duplicate action names. Feels like a good place to leave it over the break.