#1428 Data Fix For Expanded Lat/Lon in Geocoords
Fixes #1428. Updates the geocoordinate facet to generate data from the geocoordinate data that mimics the numerical summary normally returned by latitude and longitude such that filter changes are correctly reflected in those expanded facets with baseline and filtered data. Also includes some initial work for #1427 as part of making the facets interactive ties to leveraging the same existing numerical facet, but can be merged back as is without blocking further work on #1427 and without putting the front into a confusing state.