#1279 Signed SHAP weight viz
Fixes #1279: the initial pass at SHAP viz was based on absolute relative impact, and ignored the positive/negative or under/over aspect of the value regarding impact on the predicted value from the model. Now, if a SHAP weight is positive, it'll be blue, and if negative, it'll be pink. Additionally, the autoMpg set included in the repo has been updated to the min-metadata version, and the .gitignore has been updated to explicitly expect that folder to be in the repo so future updates are more easily handled as well.