#1510 Allow & Detect zip file uploads for image datasets.
Fixes #1510. Updated the file uploader to component to allow zip uploads for dataset uploads (blocking it for regression uploads as it's end point hasn't been revised as far as I know,) and updated the upload data file action to detect the file type and change the call on that basis.  It will then call importDataset right after that just like if it were a csv that had uploaded successfully.