Fix solution index number
The index number came from the list of all solutions from all requests ordered by the score.

This is not how we want to display them, so before showing them, I added a new sort to make sure they are sorted by timestamp, which is the one from their request.