#1541 Save Model and #1530 double scrollbar
Fixes #1541. Fixes #1530. Added the action and module code to post to the save route, added the button for save model and set click to trigger the save action. Adding the save button also revealed a css scope issue for the file uploader button, and while I was looking around, I quickly fixed the double scrollbar issue.