#1550 Timeseries Error Fix
Fixes #1550. Missing in data.ts & facets.ts were checks to see if we were in a view with a solution id in the route (the most reliable way to check to retrieve the solution id) and thus should be loading the predicted timeseries data in the facet for exemplars in the same way we would for other views. With that fixed, I also noticed that the prediction timeseries route was missing the URL escaping that'd already been added to the regular timeseries route, so I fixed that too.