#1583 Model Save Dialog
Fixes #1583. Updates to ResultSummaries.vue to add model save dialog with model name (mandatory) and model description (optional) with validation that model name has been filled out and updated call to the update app action for save such that model name and description are passed to the update save and solution go files which now look for modelName and description in the json payload.