adds support for display of forecasts to predictions view
SparklinePreview used the same dataset for ground truth and forecast.  In order to visualize predictions outside of train/test, we need to be able to display ground truth from the model dataset, and predictions generated from a horizon dataset.  The SparklinePreview now takes a truth dataset and a forecast dataset as arguments.  In train test scenarios, those args are the same, in prediction scenarios they will be different.