#1532 Range Mapping, Style Fixes
Handling the case where a numerical facet needs the top edge of it's range in the filter, which needs additional parse in both directions to set and get for both the URL highlight option and the selection behavior for the new facets library, as well as handling the deselection behavior in geocoordinate facets when the lat and lon facets send back null as expected on deselection. Also a style update the geocoordinate facet css which had extra padding for some reason.