#1532 - Basic Display Image Facets with new Facets
This still needs the image pattern data logic to be added to new FacetImage component; however, added FacetImage component rather than extend facet categorical for better code separation, cleaned up facet categorical for more reusability as a template for other types in future, and updated variable facets to catch and load the FacetImage component when appropriate. Still, this still needs the Pattern data logic to be added to the new FacetImage component as well.