#1680 Fix Exemplar Preload
Currently exemplar preloading subverts the natural on-demand loading of images in tables during scrolling into view or when clicking more on an image facet. This change limits that to loading the first 5 exemplars (as that's what's shown normally in image facets,) leaving the rest of the calls up to be made when additional image components are added to the page and load data.