#1809 selection fix
Fixes #1809 - Selection state has a few gotchas with the new library such that if we want to show row selection when filtering as well, we need to explicitly fill in the blanks where in row data selected is out side of the filter selection. Additionally the code needed to compare on key for the prediction facet - we don't provide error data in the table so that information can't be row highlighted, but now, everything should row highlight correctly than can be row highlighted, regardless of the filtered state.