#1839 Block Timeseries (and Geocoordinate) Creation when lacking columns
Fixes #1839. When a user goes to create a timeseries or geocoordinate but lacks the right columns to do so (a lat and lon or 2+ compatible types for geo, a value and a date or 2+ compatible types for time,) a modal appears on the join configure screen that informs the user of the issue and suggests they change some column types to fill the missing types in if possible, and then hitting ok on the modal kicks them back to their previous screen. In this I refactored some of the logic that was in variablegrouping to the types util where it could be reused in the data util as needed. new modals with computes were added to leverage the state information generated by the new data utils from the variable list.