#1755 Improve Caching To Reduce Reloading
Caching is now a dictionary of dictonaries of variable summaries keyed by variable and and call such that we can use the main list of variables to and the current route options (as a base64) to see if we've called for data previously and if so, skip loading. Further still, loading is now restricted to what's currently in the view such that we don't load anything on later pages of paginated variable lists. There's still outstanding issues with this check in, but I wanted to commit and handle conflicts with merging back master into branch before any further work.