#1755 Refactoring state out of Variable Facets
To further stress only delivering the summaries we need to the variable facet display vs. complete reloads & filtering, I've refactored the paging into a data util that takes the expected list of variables for a given variable-facet instance,  the current page number, and the count of items per page, then filters the variable array down to the expected page, then maps back the summaries for the variables given the variable name and current route key. The summaries are then passed to the summary prop for that variable-facet instance from the parent. This work needs to be done still for the results and predictions view,  but it covers the select and model views. Also removed unused deepCachable function from view actions as we handle the caching differently than when I was trying that approach, cut unneeded summary sort from mutations and related utility funciton, added getters for variable summary dictionaries, and did some other light clean up of props for readability.