#1755 Cache Training Variables in Results
Changes to the results store & view to support and leverage the same caching pattern as select and model, including some refactoring to make the data util helpers more reusable across views, as well as the some renaming of variables for greater clarity. Work should be largely the same for Predictions.