#1912 Search Fix For Select Post Refactor
Missed in the litany of things that distil expected to have everything preloaded for was facet search, so like pagination, it has to be in the routeQuery so grab it & use it to filter the variable list before updating and possibly fetching relevant summaries to the query. This first change lays the ground work for all searches in the app to populate the route while leveraging that route for the Select Target step, though I want to have the search populate from the route on refresh. Select Training is in progress, but needs even more work, then comes result and prediction views. Also cut some more unused code from the refactor.