#1901 Initial Performance Improvements
Initial work, there's more to do for results and predictions views plus clean up of unused props and logging, but added table pagination (with various layout changes & css to fit it in) to select view, using Object freeze and seal to reduce the Vue's unneeded listeners/two-way binding on things we never modify or at least don't modify deeply, using adding getters for counts rather than getting computed arrays that might have mapping steps, not searching for rowSelection data repeatedly when we can search it once and store it as we load/reload the table data into the store, directly access it when we need it for facet highlighting, saving us props after clean up. Also more router error catching, and an explanations tag we can populate with data/state in a separate story later.