#1967 Model Shouldn't Display All Features
Pulled the "top 5" from the list of variable names sent back in the model summary (filed an issue to actually send back variable scores as well so we can sort them and get a true top 5 later,) while also adding a toggle similar to dataset preview with the complete list of variables since we have the data and so people who want to look at it can see that detail before making a decision on which model to open. Also updated the dataset preview to match the cleaned up layout I set up on model preview.