#2075 Initial Lex Bar Utils
Basic utility with with helps that turn a variables into a simple lex bar language. Needs to be extended to work with all types we currently support for filtering, but now has the framework in place to extend in that fashion as well as a location for the other utility functions we need.