#2075 some style clean up
Noticed variable facets scrolling was broken in the data explorer view, so I overhauled the variable facets styling, excising as much of the custom style and div nesting as possible in favor of clean flex-boxes using bootstrap classes. This also allow for clean up some of the styling in other consumers of the variable facets component.