#2265 - backend multiple highlights
Initial Work for #2265 - changing the expectation of a single highlight to be an array in the various places needed updating (save_dataset, filter, grouped_variables, field, postgres/filter and timeseries) though those were only the places that were blocking build after the initial changes I made to the filterparam struct. it seems to work for the table endpoint now though in direct posts to it. However the front end now needs to updated to match these changes.