#2265 partial highlight removal fix.
This is better than the prior issue, where a null was set as the highlight on remove, which could in turn break highlighting all together until you renavigate to that view. However, it now clears all highlights on the removal of one from any facet, so work needs to be done to discover if we have the context on remove to only clear the target highlight, at which point, I'll probably add a optional field to clearHighlight that lets the user specify which highlights to clear and when not provided that information, clears them all.