#2265 improved highlight remove
still missing a key piece of context to make this ideal, specifically the value(s) the user is attempting to remove from the highlight, however clearHighlight was extendable to only clear a given variable rather than always all highlights, and we had the variable context at least. I need to dig more into the facets or the facet to see if there's anything already for providing the necessary context. if not, might need to extend facets.