Added check to keep from infinite looping
This is a partial fix the real problem is here https://github.com/uncharted-distil/distil/blame/main/api/model/storage/postgres/dataset.go#L858

when the ingest happens we are storing geobounds as a centroid and the extrema is calculated from it. When you only have 1 geobounds the max and min extrema are the same. Resulting in the interval for the getGeoBound buckets to be 0. When the interval is 0 it infinite loops.